+++ packages/plugins/Gfi/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## unpublished
 
+- Feature: Added optional flag `activeLayerPath`.
 - Feature: Added support for type `GeoJSON` layers.
 - Fix: Add space to dev GFI window to fully contain close button effects.